+++ examples/jaspr_basic_drag_drop/lib/main.client.dart
@@ -1,0 +1,9 @@
+library;
+
+import 'package:jaspr/client.dart';
+
+import 'app.dart';
+
+void main() {
+  runApp(const BasicDragDropApp());
+}